+++ src/test/java/org/mqttbee/mqtt/handler/ssl/SslUtilTest.java
@@ -208,7 +208,7 @@ public class SslUtilTest {
         }
 
         @Override
-        public int getHandshakeTimeoutMs() {
+        public long getHandshakeTimeoutMs() {
             return 0;
         }
     }
